test(pox-3): update tests for 2.4 activation
