+++ packages/stacking/tests/apiMockingHelpers.ts
@@ -125,28 +125,27 @@ export async function waitForCycle(cycleId: number, client?: StackingClient) {
 // === PATHS ===================================================================
 export const V2_TXS = `"0288d0bde0b0f88fad0827e35b757efec3a0cf7886c1614bfc4f81c40030a14a"`;
 export const V2_FEES = `{"estimated_cost":{"write_length":1138,"write_count":10,"read_length":60638,"read_count":34,"runtime":645443},"estimated_cost_scalar":12,"estimations":[{"fee_rate":1083.4556141387698,"fee":13001},{"fee_rate":1083.4556141387698,"fee":13001},{"fee_rate":1083.4556141387698,"fee":13001}],"cost_scalar_change_by_byte":0.00476837158203125}`;
-export const V2_POX_REGTEST = `{"contract_id":"ST000000000000000000002AMW42H.pox-2","pox_activation_threshold_ustx":600138752237389,"first_burnchain_block_height":0,"current_burnchain_block_height":8249,"prepare_phase_block_length":1,"reward_phase_block_length":4,"reward_slots":8,"rejection_fraction":3333333333333333,"total_liquid_supply_ustx":60013875223738910,"current_cycle":{"id":1649,"min_threshold_ustx":1875440000000000,"stacked_ustx":0,"is_pox_active":false},"next_cycle":{"id":1650,"min_threshold_ustx":1875440000000000,"min_increment_ustx":7501734402967,"stacked_ustx":0,"prepare_phase_start_block_height":8249,"blocks_until_prepare_phase":0,"reward_phase_start_block_height":8250,"blocks_until_reward_phase":1,"ustx_until_pox_rejection":1589413198036825600},"min_amount_ustx":1875440000000000,"prepare_cycle_length":1,"reward_cycle_id":1649,"reward_cycle_length":5,"rejection_votes_left_required":1589413198036825600,"next_reward_cycle_in":1,"contract_versions":[{"contract_id":"ST000000000000000000002AMW42H.pox","activation_burnchain_block_height":0,"first_reward_cycle_id":0},{"contract_id":"ST000000000000000000002AMW42H.pox-2","activation_burnchain_block_height":104,"first_reward_cycle_id":21}]}`;
-export const V2_POX_STACKS_TEL = `{"contract_id":"ST000000000000000000002AMW42H.pox-2","pox_activation_threshold_ustx":600666564165569,"first_burnchain_block_height":0,"current_burnchain_block_height":60989,"prepare_phase_block_length":1,"reward_phase_block_length":4,"reward_slots":8,"rejection_fraction":3333333333333333,"total_liquid_supply_ustx":60066656416556970,"current_cycle":{"id":12197,"min_threshold_ustx":1877090000000000,"stacked_ustx":0,"is_pox_active":false},"next_cycle":{"id":12198,"min_threshold_ustx":1877090000000000,"min_increment_ustx":7508332052069,"stacked_ustx":0,"prepare_phase_start_block_height":60989,"blocks_until_prepare_phase":0,"reward_phase_start_block_height":60990,"blocks_until_reward_phase":1,"ustx_until_pox_rejection":11416974460158247000},"min_amount_ustx":1877090000000000,"prepare_cycle_length":1,"reward_cycle_id":12197,"reward_cycle_length":5,"rejection_votes_left_required":11416974460158247000,"next_reward_cycle_in":1,"contract_versions":[{"contract_id":"ST000000000000000000002AMW42H.pox","activation_burnchain_block_height":0,"first_reward_cycle_id":0},{"contract_id":"ST000000000000000000002AMW42H.pox-2","activation_burnchain_block_height":104,"first_reward_cycle_id":21}]}`;
-export const V2_POX_INTERFACE_POX_2 = `{"functions":[{"name":"add-pox-addr-to-ith-reward-cycle","access":"private","args":[{"name":"cycle-index","type":"uint128"},{"name":"params","type":{"tuple":[{"name":"amount-ustx","type":"uint128"},{"name":"first-reward-cycle","type":"uint128"},{"name":"i","type":"uint128"},{"name":"num-cycles","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-set-indexes","type":{"list":{"type":"uint128","length":12}}},{"name":"stacker","type":{"optional":"principal"}}]}}],"outputs":{"type":{"tuple":[{"name":"amount-ustx","type":"uint128"},{"name":"first-reward-cycle","type":"uint128"},{"name":"i","type":"uint128"},{"name":"num-cycles","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-set-indexes","type":{"list":{"type":"uint128","length":12}}},{"name":"stacker","type":{"optional":"principal"}}]}}},{"name":"add-pox-addr-to-reward-cycles","access":"private","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"first-reward-cycle","type":"uint128"},{"name":"num-cycles","type":"uint128"},{"name":"amount-ustx","type":"uint128"},{"name":"stacker","type":"principal"}],"outputs":{"type":{"response":{"ok":{"list":{"type":"uint128","length":12}},"error":"int128"}}}},{"name":"add-pox-partial-stacked","access":"private","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"first-reward-cycle","type":"uint128"},{"name":"num-cycles","type":"uint128"},{"name":"amount-ustx","type":"uint128"}],"outputs":{"type":"bool"}},{"name":"add-pox-partial-stacked-to-ith-cycle","access":"private","args":[{"name":"cycle-index","type":"uint128"},{"name":"params","type":{"tuple":[{"name":"amount-ustx","type":"uint128"},{"name":"num-cycles","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"}]}}],"outputs":{"type":{"tuple":[{"name":"amount-ustx","type":"uint128"},{"name":"num-cycles","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"}]}}},{"name":"append-reward-cycle-pox-addr","access":"private","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"},{"name":"amount-ustx","type":"uint128"},{"name":"stacker","type":{"optional":"principal"}}],"outputs":{"type":"uint128"}},{"name":"fold-unlock-reward-cycle","access":"private","args":[{"name":"set-index","type":"uint128"},{"name":"data-res","type":{"response":{"ok":{"tuple":[{"name":"cycle","type":"uint128"},{"name":"first-unlocked-cycle","type":"uint128"},{"name":"stacker","type":"principal"}]},"error":"int128"}}}],"outputs":{"type":{"response":{"ok":{"tuple":[{"name":"cycle","type":"uint128"},{"name":"first-unlocked-cycle","type":"uint128"},{"name":"stacker","type":"principal"}]},"error":"int128"}}}},{"name":"handle-unlock","access":"private","args":[{"name":"user","type":"principal"},{"name":"amount-locked","type":"uint128"},{"name":"cycle-to-unlock","type":"uint128"}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"increase-reward-cycle-entry","access":"private","args":[{"name":"reward-cycle-index","type":"uint128"},{"name":"updates","type":{"optional":{"tuple":[{"name":"add-amount","type":"uint128"},{"name":"first-cycle","type":"uint128"},{"name":"reward-cycle","type":"uint128"},{"name":"stacker","type":"principal"}]}}}],"outputs":{"type":{"optional":{"tuple":[{"name":"add-amount","type":"uint128"},{"name":"first-cycle","type":"uint128"},{"name":"reward-cycle","type":"uint128"},{"name":"stacker","type":"principal"}]}}}},{"name":"inner-stack-aggregation-commit","access":"private","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":{"response":{"ok":"uint128","error":"int128"}}}},{"name":"allow-contract-caller","access":"public","args":[{"name":"caller","type":"principal"},{"name":"until-burn-ht","type":{"optional":"uint128"}}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"delegate-stack-extend","access":"public","args":[{"name":"stacker","type":"principal"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"extend-count","type":"uint128"}],"outputs":{"type":{"response":{"ok":{"tuple":[{"name":"stacker","type":"principal"},{"name":"unlock-burn-height","type":"uint128"}]},"error":"int128"}}}},{"name":"delegate-stack-increase","access":"public","args":[{"name":"stacker","type":"principal"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"increase-by","type":"uint128"}],"outputs":{"type":{"response":{"ok":{"tuple":[{"name":"stacker","type":"principal"},{"name":"total-locked","type":"uint128"}]},"error":"int128"}}}},{"name":"delegate-stack-stx","access":"public","args":[{"name":"stacker","type":"principal"},{"name":"amount-ustx","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"start-burn-ht","type":"uint128"},{"name":"lock-period","type":"uint128"}],"outputs":{"type":{"response":{"ok":{"tuple":[{"name":"lock-amount","type":"uint128"},{"name":"stacker","type":"principal"},{"name":"unlock-burn-height","type":"uint128"}]},"error":"int128"}}}},{"name":"delegate-stx","access":"public","args":[{"name":"amount-ustx","type":"uint128"},{"name":"delegate-to","type":"principal"},{"name":"until-burn-ht","type":{"optional":"uint128"}},{"name":"pox-addr","type":{"optional":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}}}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"disallow-contract-caller","access":"public","args":[{"name":"caller","type":"principal"}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"reject-pox","access":"public","args":[],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"revoke-delegate-stx","access":"public","args":[],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"set-burnchain-parameters","access":"public","args":[{"name":"first-burn-height","type":"uint128"},{"name":"prepare-cycle-length","type":"uint128"},{"name":"reward-cycle-length","type":"uint128"},{"name":"rejection-fraction","type":"uint128"},{"name":"begin-2-1-reward-cycle","type":"uint128"}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"stack-aggregation-commit","access":"public","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"stack-aggregation-commit-indexed","access":"public","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":{"response":{"ok":"uint128","error":"int128"}}}},{"name":"stack-aggregation-increase","access":"public","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"},{"name":"reward-cycle-index","type":"uint128"}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"stack-extend","access":"public","args":[{"name":"extend-count","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}}],"outputs":{"type":{"response":{"ok":{"tuple":[{"name":"stacker","type":"principal"},{"name":"unlock-burn-height","type":"uint128"}]},"error":"int128"}}}},{"name":"stack-increase","access":"public","args":[{"name":"increase-by","type":"uint128"}],"outputs":{"type":{"response":{"ok":{"tuple":[{"name":"stacker","type":"principal"},{"name":"total-locked","type":"uint128"}]},"error":"int128"}}}},{"name":"stack-stx","access":"public","args":[{"name":"amount-ustx","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"start-burn-ht","type":"uint128"},{"name":"lock-period","type":"uint128"}],"outputs":{"type":{"response":{"ok":{"tuple":[{"name":"lock-amount","type":"uint128"},{"name":"stacker","type":"principal"},{"name":"unlock-burn-height","type":"uint128"}]},"error":"int128"}}}},{"name":"burn-height-to-reward-cycle","access":"read_only","args":[{"name":"height","type":"uint128"}],"outputs":{"type":"uint128"}},{"name":"can-stack-stx","access":"read_only","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"amount-ustx","type":"uint128"},{"name":"first-reward-cycle","type":"uint128"},{"name":"num-cycles","type":"uint128"}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"check-caller-allowed","access":"read_only","args":[],"outputs":{"type":"bool"}},{"name":"check-pox-addr-hashbytes","access":"read_only","args":[{"name":"version","type":{"buffer":{"length":1}}},{"name":"hashbytes","type":{"buffer":{"length":32}}}],"outputs":{"type":"bool"}},{"name":"check-pox-addr-version","access":"read_only","args":[{"name":"version","type":{"buffer":{"length":1}}}],"outputs":{"type":"bool"}},{"name":"check-pox-lock-period","access":"read_only","args":[{"name":"lock-period","type":"uint128"}],"outputs":{"type":"bool"}},{"name":"current-pox-reward-cycle","access":"read_only","args":[],"outputs":{"type":"uint128"}},{"name":"get-allowance-contract-callers","access":"read_only","args":[{"name":"sender","type":"principal"},{"name":"calling-contract","type":"principal"}],"outputs":{"type":{"optional":{"tuple":[{"name":"until-burn-ht","type":{"optional":"uint128"}}]}}}},{"name":"get-check-delegation","access":"read_only","args":[{"name":"stacker","type":"principal"}],"outputs":{"type":{"optional":{"tuple":[{"name":"amount-ustx","type":"uint128"},{"name":"delegated-to","type":"principal"},{"name":"pox-addr","type":{"optional":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}}},{"name":"until-burn-ht","type":{"optional":"uint128"}}]}}}},{"name":"get-delegation-info","access":"read_only","args":[{"name":"stacker","type":"principal"}],"outputs":{"type":{"optional":{"tuple":[{"name":"amount-ustx","type":"uint128"},{"name":"delegated-to","type":"principal"},{"name":"pox-addr","type":{"optional":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}}},{"name":"until-burn-ht","type":{"optional":"uint128"}}]}}}},{"name":"get-num-reward-set-pox-addresses","access":"read_only","args":[{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":"uint128"}},{"name":"get-partial-stacked-by-cycle","access":"read_only","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"},{"name":"sender","type":"principal"}],"outputs":{"type":{"optional":{"tuple":[{"name":"stacked-amount","type":"uint128"}]}}}},{"name":"get-pox-info","access":"read_only","args":[],"outputs":{"type":{"response":{"ok":{"tuple":[{"name":"current-rejection-votes","type":"uint128"},{"name":"first-burnchain-block-height","type":"uint128"},{"name":"min-amount-ustx","type":"uint128"},{"name":"prepare-cycle-length","type":"uint128"},{"name":"rejection-fraction","type":"uint128"},{"name":"reward-cycle-id","type":"uint128"},{"name":"reward-cycle-length","type":"uint128"},{"name":"total-liquid-supply-ustx","type":"uint128"}]},"error":"none"}}}},{"name":"get-pox-rejection","access":"read_only","args":[{"name":"stacker","type":"principal"},{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":{"optional":{"tuple":[{"name":"amount","type":"uint128"}]}}}},{"name":"get-reward-set-pox-address","access":"read_only","args":[{"name":"reward-cycle","type":"uint128"},{"name":"index","type":"uint128"}],"outputs":{"type":{"optional":{"tuple":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"stacker","type":{"optional":"principal"}},{"name":"total-ustx","type":"uint128"}]}}}},{"name":"get-reward-set-size","access":"read_only","args":[{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":"uint128"}},{"name":"get-stacker-info","access":"read_only","args":[{"name":"stacker","type":"principal"}],"outputs":{"type":{"optional":{"tuple":[{"name":"first-reward-cycle","type":"uint128"},{"name":"lock-period","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-set-indexes","type":{"list":{"type":"uint128","length":12}}}]}}}},{"name":"get-stacking-minimum","access":"read_only","args":[],"outputs":{"type":"uint128"}},{"name":"get-total-pox-rejection","access":"read_only","args":[{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":"uint128"}},{"name":"get-total-ustx-stacked","access":"read_only","args":[{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":"uint128"}},{"name":"is-pox-active","access":"read_only","args":[{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":"bool"}},{"name":"minimal-can-stack-stx","access":"read_only","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"amount-ustx","type":"uint128"},{"name":"first-reward-cycle","type":"uint128"},{"name":"num-cycles","type":"uint128"}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"next-cycle-rejection-votes","access":"read_only","args":[],"outputs":{"type":"uint128"}},{"name":"reward-cycle-to-burn-height","access":"read_only","args":[{"name":"cycle","type":"uint128"}],"outputs":{"type":"uint128"}}],"variables":[{"name":"ADDRESS_VERSION_NATIVE_P2TR","type":{"buffer":{"length":1}},"access":"constant"},{"name":"ADDRESS_VERSION_NATIVE_P2WPKH","type":{"buffer":{"length":1}},"access":"constant"},{"name":"ADDRESS_VERSION_NATIVE_P2WSH","type":{"buffer":{"length":1}},"access":"constant"},{"name":"ADDRESS_VERSION_P2PKH","type":{"buffer":{"length":1}},"access":"constant"},{"name":"ADDRESS_VERSION_P2SH","type":{"buffer":{"length":1}},"access":"constant"},{"name":"ADDRESS_VERSION_P2WPKH","type":{"buffer":{"length":1}},"access":"constant"},{"name":"ADDRESS_VERSION_P2WSH","type":{"buffer":{"length":1}},"access":"constant"},{"name":"ERR_DELEGATION_EXPIRES_DURING_LOCK","type":"int128","access":"constant"},{"name":"ERR_DELEGATION_NO_REWARD_SLOT","type":"int128","access":"constant"},{"name":"ERR_DELEGATION_POX_ADDR_REQUIRED","type":"int128","access":"constant"},{"name":"ERR_DELEGATION_TOO_MUCH_LOCKED","type":"int128","access":"constant"},{"name":"ERR_DELEGATION_WRONG_REWARD_SLOT","type":"int128","access":"constant"},{"name":"ERR_INVALID_START_BURN_HEIGHT","type":"int128","access":"constant"},{"name":"ERR_NOT_ALLOWED","type":"int128","access":"constant"},{"name":"ERR_NOT_CURRENT_STACKER","type":"int128","access":"constant"},{"name":"ERR_STACKING_ALREADY_DELEGATED","type":"int128","access":"constant"},{"name":"ERR_STACKING_ALREADY_REJECTED","type":"int128","access":"constant"},{"name":"ERR_STACKING_ALREADY_STACKED","type":"int128","access":"constant"},{"name":"ERR_STACKING_CORRUPTED_STATE","type":"int128","access":"constant"},{"name":"ERR_STACKING_EXPIRED","type":"int128","access":"constant"},{"name":"ERR_STACKING_INSUFFICIENT_FUNDS","type":"int128","access":"constant"},{"name":"ERR_STACKING_INVALID_AMOUNT","type":"int128","access":"constant"},{"name":"ERR_STACKING_INVALID_LOCK_PERIOD","type":"int128","access":"constant"},{"name":"ERR_STACKING_INVALID_POX_ADDRESS","type":"int128","access":"constant"},{"name":"ERR_STACKING_NO_SUCH_PRINCIPAL","type":"int128","access":"constant"},{"name":"ERR_STACKING_PERMISSION_DENIED","type":"int128","access":"constant"},{"name":"ERR_STACKING_POX_ADDRESS_IN_USE","type":"int128","access":"constant"},{"name":"ERR_STACKING_STX_LOCKED","type":"int128","access":"constant"},{"name":"ERR_STACKING_THRESHOLD_NOT_MET","type":"int128","access":"constant"},{"name":"ERR_STACKING_UNREACHABLE","type":"int128","access":"constant"},{"name":"ERR_STACK_EXTEND_NOT_LOCKED","type":"int128","access":"constant"},{"name":"ERR_STACK_INCREASE_NOT_LOCKED","type":"int128","access":"constant"},{"name":"MAX_ADDRESS_VERSION","type":"uint128","access":"constant"},{"name":"MAX_ADDRESS_VERSION_BUFF_20","type":"uint128","access":"constant"},{"name":"MAX_ADDRESS_VERSION_BUFF_32","type":"uint128","access":"constant"},{"name":"MAX_POX_REWARD_CYCLES","type":"uint128","access":"constant"},{"name":"MIN_POX_REWARD_CYCLES","type":"uint128","access":"constant"},{"name":"POX_REJECTION_FRACTION","type":"uint128","access":"constant"},{"name":"PREPARE_CYCLE_LENGTH","type":"uint128","access":"constant"},{"name":"REWARD_CYCLE_LENGTH","type":"uint128","access":"constant"},{"name":"STACKING_THRESHOLD_100","type":"uint128","access":"constant"},{"name":"STACKING_THRESHOLD_25","type":"uint128","access":"constant"},{"name":"configured","type":"bool","access":"variable"},{"name":"first-2-1-reward-cycle","type":"uint128","access":"variable"},{"name":"first-burnchain-block-height","type":"uint128","access":"variable"},{"name":"pox-prepare-cycle-length","type":"uint128","access":"variable"},{"name":"pox-rejection-fraction","type":"uint128","access":"variable"},{"name":"pox-reward-cycle-length","type":"uint128","access":"variable"}],"maps":[{"name":"allowance-contract-callers","key":{"tuple":[{"name":"contract-caller","type":"principal"},{"name":"sender","type":"principal"}]},"value":{"tuple":[{"name":"until-burn-ht","type":{"optional":"uint128"}}]}},{"name":"delegation-state","key":{"tuple":[{"name":"stacker","type":"principal"}]},"value":{"tuple":[{"name":"amount-ustx","type":"uint128"},{"name":"delegated-to","type":"principal"},{"name":"pox-addr","type":{"optional":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}}},{"name":"until-burn-ht","type":{"optional":"uint128"}}]}},{"name":"partial-stacked-by-cycle","key":{"tuple":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"},{"name":"sender","type":"principal"}]},"value":{"tuple":[{"name":"stacked-amount","type":"uint128"}]}},{"name":"reward-cycle-pox-address-list","key":{"tuple":[{"name":"index","type":"uint128"},{"name":"reward-cycle","type":"uint128"}]},"value":{"tuple":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"stacker","type":{"optional":"principal"}},{"name":"total-ustx","type":"uint128"}]}},{"name":"reward-cycle-pox-address-list-len","key":{"tuple":[{"name":"reward-cycle","type":"uint128"}]},"value":{"tuple":[{"name":"len","type":"uint128"}]}},{"name":"reward-cycle-total-stacked","key":{"tuple":[{"name":"reward-cycle","type":"uint128"}]},"value":{"tuple":[{"name":"total-ustx","type":"uint128"}]}},{"name":"stacking-rejection","key":{"tuple":[{"name":"reward-cycle","type":"uint128"}]},"value":{"tuple":[{"name":"amount","type":"uint128"}]}},{"name":"stacking-rejectors","key":{"tuple":[{"name":"reward-cycle","type":"uint128"},{"name":"stacker","type":"principal"}]},"value":{"tuple":[{"name":"amount","type":"uint128"}]}},{"name":"stacking-state","key":{"tuple":[{"name":"stacker","type":"principal"}]},"value":{"tuple":[{"name":"first-reward-cycle","type":"uint128"},{"name":"lock-period","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-set-indexes","type":{"list":{"type":"uint128","length":12}}}]}}],"fungible_tokens":[],"non_fungible_tokens":[],"clarity_version":"Clarity2"}`;
+export const V2_POX_MAINNET_POX_2 = `{"contract_id":"SP000000000000000000002Q6VF78.pox-2","pox_activation_threshold_ustx":68809653919448,"first_burnchain_block_height":666050,"current_burnchain_block_height":788791,"prepare_phase_block_length":100,"reward_phase_block_length":2000,"reward_slots":4000,"rejection_fraction":25,"total_liquid_supply_ustx":1376193078388960,"current_cycle":{"id":58,"min_threshold_ustx":90000000000,"stacked_ustx":225878587321503,"is_pox_active":false},"next_cycle":{"id":59,"min_threshold_ustx":90000000000,"min_increment_ustx":68809653919,"stacked_ustx":195129336384433,"prepare_phase_start_block_height":789850,"blocks_until_prepare_phase":1059,"reward_phase_start_block_height":789950,"blocks_until_reward_phase":1159,"ustx_until_pox_rejection":344048269597225},"min_amount_ustx":90000000000,"prepare_cycle_length":100,"reward_cycle_id":58,"reward_cycle_length":2100,"rejection_votes_left_required":344048269597225,"next_reward_cycle_in":1159,"contract_versions":[{"contract_id":"SP000000000000000000002Q6VF78.pox","activation_burnchain_block_height":666050,"first_reward_cycle_id":0},{"contract_id":"SP000000000000000000002Q6VF78.pox-2","activation_burnchain_block_height":781552,"first_reward_cycle_id":56}]}`;
+export const V2_POX_REGTEST_POX_3 = `{"contract_id":"ST000000000000000000002AMW42H.pox-3","pox_activation_threshold_ustx":600059459544055,"first_burnchain_block_height":0,"current_burnchain_block_height":408,"prepare_phase_block_length":1,"reward_phase_block_length":4,"reward_slots":8,"rejection_fraction":3333333333333333,"total_liquid_supply_ustx":60005945954405576,"current_cycle":{"id":81,"min_threshold_ustx":1875190000000000,"stacked_ustx":1875190000000000,"is_pox_active":false},"next_cycle":{"id":82,"min_threshold_ustx":1875190000000000,"min_increment_ustx":7500743244300,"stacked_ustx":1875190000000000,"prepare_phase_start_block_height":409,"blocks_until_prepare_phase":1,"reward_phase_start_block_height":410,"blocks_until_reward_phase":2,"ustx_until_pox_rejection":12734475822413576000},"min_amount_ustx":1875190000000000,"prepare_cycle_length":1,"reward_cycle_id":81,"reward_cycle_length":5,"rejection_votes_left_required":12734475822413576000,"next_reward_cycle_in":2,"contract_versions":[{"contract_id":"ST000000000000000000002AMW42H.pox","activation_burnchain_block_height":0,"first_reward_cycle_id":0},{"contract_id":"ST000000000000000000002AMW42H.pox-2","activation_burnchain_block_height":107,"first_reward_cycle_id":22},{"contract_id":"ST000000000000000000002AMW42H.pox-3","activation_burnchain_block_height":111,"first_reward_cycle_id":23}]}`;
+export const V2_POX_INTERFACE_POX_3 = `{"functions":[{"name":"add-pox-addr-to-ith-reward-cycle","access":"private","args":[{"name":"cycle-index","type":"uint128"},{"name":"params","type":{"tuple":[{"name":"amount-ustx","type":"uint128"},{"name":"first-reward-cycle","type":"uint128"},{"name":"i","type":"uint128"},{"name":"num-cycles","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-set-indexes","type":{"list":{"type":"uint128","length":12}}},{"name":"stacker","type":{"optional":"principal"}}]}}],"outputs":{"type":{"tuple":[{"name":"amount-ustx","type":"uint128"},{"name":"first-reward-cycle","type":"uint128"},{"name":"i","type":"uint128"},{"name":"num-cycles","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-set-indexes","type":{"list":{"type":"uint128","length":12}}},{"name":"stacker","type":{"optional":"principal"}}]}}},{"name":"add-pox-addr-to-reward-cycles","access":"private","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"first-reward-cycle","type":"uint128"},{"name":"num-cycles","type":"uint128"},{"name":"amount-ustx","type":"uint128"},{"name":"stacker","type":"principal"}],"outputs":{"type":{"response":{"ok":{"list":{"type":"uint128","length":12}},"error":"int128"}}}},{"name":"add-pox-partial-stacked","access":"private","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"first-reward-cycle","type":"uint128"},{"name":"num-cycles","type":"uint128"},{"name":"amount-ustx","type":"uint128"}],"outputs":{"type":"bool"}},{"name":"add-pox-partial-stacked-to-ith-cycle","access":"private","args":[{"name":"cycle-index","type":"uint128"},{"name":"params","type":{"tuple":[{"name":"amount-ustx","type":"uint128"},{"name":"num-cycles","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"}]}}],"outputs":{"type":{"tuple":[{"name":"amount-ustx","type":"uint128"},{"name":"num-cycles","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"}]}}},{"name":"append-reward-cycle-pox-addr","access":"private","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"},{"name":"amount-ustx","type":"uint128"},{"name":"stacker","type":{"optional":"principal"}}],"outputs":{"type":"uint128"}},{"name":"fold-unlock-reward-cycle","access":"private","args":[{"name":"set-index","type":"uint128"},{"name":"data-res","type":{"response":{"ok":{"tuple":[{"name":"cycle","type":"uint128"},{"name":"first-unlocked-cycle","type":"uint128"},{"name":"stacker","type":"principal"}]},"error":"int128"}}}],"outputs":{"type":{"response":{"ok":{"tuple":[{"name":"cycle","type":"uint128"},{"name":"first-unlocked-cycle","type":"uint128"},{"name":"stacker","type":"principal"}]},"error":"int128"}}}},{"name":"handle-unlock","access":"private","args":[{"name":"user","type":"principal"},{"name":"amount-locked","type":"uint128"},{"name":"cycle-to-unlock","type":"uint128"}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"increase-reward-cycle-entry","access":"private","args":[{"name":"reward-cycle-index","type":"uint128"},{"name":"updates","type":{"optional":{"tuple":[{"name":"add-amount","type":"uint128"},{"name":"first-cycle","type":"uint128"},{"name":"reward-cycle","type":"uint128"},{"name":"stacker","type":"principal"}]}}}],"outputs":{"type":{"optional":{"tuple":[{"name":"add-amount","type":"uint128"},{"name":"first-cycle","type":"uint128"},{"name":"reward-cycle","type":"uint128"},{"name":"stacker","type":"principal"}]}}}},{"name":"inner-stack-aggregation-commit","access":"private","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":{"response":{"ok":"uint128","error":"int128"}}}},{"name":"allow-contract-caller","access":"public","args":[{"name":"caller","type":"principal"},{"name":"until-burn-ht","type":{"optional":"uint128"}}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"delegate-stack-extend","access":"public","args":[{"name":"stacker","type":"principal"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"extend-count","type":"uint128"}],"outputs":{"type":{"response":{"ok":{"tuple":[{"name":"stacker","type":"principal"},{"name":"unlock-burn-height","type":"uint128"}]},"error":"int128"}}}},{"name":"delegate-stack-increase","access":"public","args":[{"name":"stacker","type":"principal"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"increase-by","type":"uint128"}],"outputs":{"type":{"response":{"ok":{"tuple":[{"name":"stacker","type":"principal"},{"name":"total-locked","type":"uint128"}]},"error":"int128"}}}},{"name":"delegate-stack-stx","access":"public","args":[{"name":"stacker","type":"principal"},{"name":"amount-ustx","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"start-burn-ht","type":"uint128"},{"name":"lock-period","type":"uint128"}],"outputs":{"type":{"response":{"ok":{"tuple":[{"name":"lock-amount","type":"uint128"},{"name":"stacker","type":"principal"},{"name":"unlock-burn-height","type":"uint128"}]},"error":"int128"}}}},{"name":"delegate-stx","access":"public","args":[{"name":"amount-ustx","type":"uint128"},{"name":"delegate-to","type":"principal"},{"name":"until-burn-ht","type":{"optional":"uint128"}},{"name":"pox-addr","type":{"optional":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}}}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"disallow-contract-caller","access":"public","args":[{"name":"caller","type":"principal"}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"reject-pox","access":"public","args":[],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"revoke-delegate-stx","access":"public","args":[],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"set-burnchain-parameters","access":"public","args":[{"name":"first-burn-height","type":"uint128"},{"name":"prepare-cycle-length","type":"uint128"},{"name":"reward-cycle-length","type":"uint128"},{"name":"rejection-fraction","type":"uint128"},{"name":"begin-2-1-reward-cycle","type":"uint128"}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"stack-aggregation-commit","access":"public","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"stack-aggregation-commit-indexed","access":"public","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":{"response":{"ok":"uint128","error":"int128"}}}},{"name":"stack-aggregation-increase","access":"public","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"},{"name":"reward-cycle-index","type":"uint128"}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"stack-extend","access":"public","args":[{"name":"extend-count","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}}],"outputs":{"type":{"response":{"ok":{"tuple":[{"name":"stacker","type":"principal"},{"name":"unlock-burn-height","type":"uint128"}]},"error":"int128"}}}},{"name":"stack-increase","access":"public","args":[{"name":"increase-by","type":"uint128"}],"outputs":{"type":{"response":{"ok":{"tuple":[{"name":"stacker","type":"principal"},{"name":"total-locked","type":"uint128"}]},"error":"int128"}}}},{"name":"stack-stx","access":"public","args":[{"name":"amount-ustx","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"start-burn-ht","type":"uint128"},{"name":"lock-period","type":"uint128"}],"outputs":{"type":{"response":{"ok":{"tuple":[{"name":"lock-amount","type":"uint128"},{"name":"stacker","type":"principal"},{"name":"unlock-burn-height","type":"uint128"}]},"error":"int128"}}}},{"name":"burn-height-to-reward-cycle","access":"read_only","args":[{"name":"height","type":"uint128"}],"outputs":{"type":"uint128"}},{"name":"can-stack-stx","access":"read_only","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"amount-ustx","type":"uint128"},{"name":"first-reward-cycle","type":"uint128"},{"name":"num-cycles","type":"uint128"}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"check-caller-allowed","access":"read_only","args":[],"outputs":{"type":"bool"}},{"name":"check-pox-addr-hashbytes","access":"read_only","args":[{"name":"version","type":{"buffer":{"length":1}}},{"name":"hashbytes","type":{"buffer":{"length":32}}}],"outputs":{"type":"bool"}},{"name":"check-pox-addr-version","access":"read_only","args":[{"name":"version","type":{"buffer":{"length":1}}}],"outputs":{"type":"bool"}},{"name":"check-pox-lock-period","access":"read_only","args":[{"name":"lock-period","type":"uint128"}],"outputs":{"type":"bool"}},{"name":"current-pox-reward-cycle","access":"read_only","args":[],"outputs":{"type":"uint128"}},{"name":"get-allowance-contract-callers","access":"read_only","args":[{"name":"sender","type":"principal"},{"name":"calling-contract","type":"principal"}],"outputs":{"type":{"optional":{"tuple":[{"name":"until-burn-ht","type":{"optional":"uint128"}}]}}}},{"name":"get-check-delegation","access":"read_only","args":[{"name":"stacker","type":"principal"}],"outputs":{"type":{"optional":{"tuple":[{"name":"amount-ustx","type":"uint128"},{"name":"delegated-to","type":"principal"},{"name":"pox-addr","type":{"optional":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}}},{"name":"until-burn-ht","type":{"optional":"uint128"}}]}}}},{"name":"get-delegation-info","access":"read_only","args":[{"name":"stacker","type":"principal"}],"outputs":{"type":{"optional":{"tuple":[{"name":"amount-ustx","type":"uint128"},{"name":"delegated-to","type":"principal"},{"name":"pox-addr","type":{"optional":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}}},{"name":"until-burn-ht","type":{"optional":"uint128"}}]}}}},{"name":"get-num-reward-set-pox-addresses","access":"read_only","args":[{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":"uint128"}},{"name":"get-partial-stacked-by-cycle","access":"read_only","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"},{"name":"sender","type":"principal"}],"outputs":{"type":{"optional":{"tuple":[{"name":"stacked-amount","type":"uint128"}]}}}},{"name":"get-pox-info","access":"read_only","args":[],"outputs":{"type":{"response":{"ok":{"tuple":[{"name":"current-rejection-votes","type":"uint128"},{"name":"first-burnchain-block-height","type":"uint128"},{"name":"min-amount-ustx","type":"uint128"},{"name":"prepare-cycle-length","type":"uint128"},{"name":"rejection-fraction","type":"uint128"},{"name":"reward-cycle-id","type":"uint128"},{"name":"reward-cycle-length","type":"uint128"},{"name":"total-liquid-supply-ustx","type":"uint128"}]},"error":"none"}}}},{"name":"get-pox-rejection","access":"read_only","args":[{"name":"stacker","type":"principal"},{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":{"optional":{"tuple":[{"name":"amount","type":"uint128"}]}}}},{"name":"get-reward-set-pox-address","access":"read_only","args":[{"name":"reward-cycle","type":"uint128"},{"name":"index","type":"uint128"}],"outputs":{"type":{"optional":{"tuple":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"stacker","type":{"optional":"principal"}},{"name":"total-ustx","type":"uint128"}]}}}},{"name":"get-reward-set-size","access":"read_only","args":[{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":"uint128"}},{"name":"get-stacker-info","access":"read_only","args":[{"name":"stacker","type":"principal"}],"outputs":{"type":{"optional":{"tuple":[{"name":"delegated-to","type":{"optional":"principal"}},{"name":"first-reward-cycle","type":"uint128"},{"name":"lock-period","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-set-indexes","type":{"list":{"type":"uint128","length":12}}}]}}}},{"name":"get-stacking-minimum","access":"read_only","args":[],"outputs":{"type":"uint128"}},{"name":"get-total-pox-rejection","access":"read_only","args":[{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":"uint128"}},{"name":"get-total-ustx-stacked","access":"read_only","args":[{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":"uint128"}},{"name":"is-pox-active","access":"read_only","args":[{"name":"reward-cycle","type":"uint128"}],"outputs":{"type":"bool"}},{"name":"minimal-can-stack-stx","access":"read_only","args":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"amount-ustx","type":"uint128"},{"name":"first-reward-cycle","type":"uint128"},{"name":"num-cycles","type":"uint128"}],"outputs":{"type":{"response":{"ok":"bool","error":"int128"}}}},{"name":"next-cycle-rejection-votes","access":"read_only","args":[],"outputs":{"type":"uint128"}},{"name":"reward-cycle-to-burn-height","access":"read_only","args":[{"name":"cycle","type":"uint128"}],"outputs":{"type":"uint128"}}],"variables":[{"name":"ADDRESS_VERSION_NATIVE_P2TR","type":{"buffer":{"length":1}},"access":"constant"},{"name":"ADDRESS_VERSION_NATIVE_P2WPKH","type":{"buffer":{"length":1}},"access":"constant"},{"name":"ADDRESS_VERSION_NATIVE_P2WSH","type":{"buffer":{"length":1}},"access":"constant"},{"name":"ADDRESS_VERSION_P2PKH","type":{"buffer":{"length":1}},"access":"constant"},{"name":"ADDRESS_VERSION_P2SH","type":{"buffer":{"length":1}},"access":"constant"},{"name":"ADDRESS_VERSION_P2WPKH","type":{"buffer":{"length":1}},"access":"constant"},{"name":"ADDRESS_VERSION_P2WSH","type":{"buffer":{"length":1}},"access":"constant"},{"name":"ERR_DELEGATION_EXPIRES_DURING_LOCK","type":"int128","access":"constant"},{"name":"ERR_DELEGATION_NO_REWARD_SLOT","type":"int128","access":"constant"},{"name":"ERR_DELEGATION_POX_ADDR_REQUIRED","type":"int128","access":"constant"},{"name":"ERR_DELEGATION_TOO_MUCH_LOCKED","type":"int128","access":"constant"},{"name":"ERR_DELEGATION_WRONG_REWARD_SLOT","type":"int128","access":"constant"},{"name":"ERR_INVALID_START_BURN_HEIGHT","type":"int128","access":"constant"},{"name":"ERR_NOT_ALLOWED","type":"int128","access":"constant"},{"name":"ERR_NOT_CURRENT_STACKER","type":"int128","access":"constant"},{"name":"ERR_STACKING_ALREADY_DELEGATED","type":"int128","access":"constant"},{"name":"ERR_STACKING_ALREADY_REJECTED","type":"int128","access":"constant"},{"name":"ERR_STACKING_ALREADY_STACKED","type":"int128","access":"constant"},{"name":"ERR_STACKING_CORRUPTED_STATE","type":"int128","access":"constant"},{"name":"ERR_STACKING_EXPIRED","type":"int128","access":"constant"},{"name":"ERR_STACKING_INSUFFICIENT_FUNDS","type":"int128","access":"constant"},{"name":"ERR_STACKING_INVALID_AMOUNT","type":"int128","access":"constant"},{"name":"ERR_STACKING_INVALID_LOCK_PERIOD","type":"int128","access":"constant"},{"name":"ERR_STACKING_INVALID_POX_ADDRESS","type":"int128","access":"constant"},{"name":"ERR_STACKING_IS_DELEGATED","type":"int128","access":"constant"},{"name":"ERR_STACKING_NOT_DELEGATED","type":"int128","access":"constant"},{"name":"ERR_STACKING_NO_SUCH_PRINCIPAL","type":"int128","access":"constant"},{"name":"ERR_STACKING_PERMISSION_DENIED","type":"int128","access":"constant"},{"name":"ERR_STACKING_POX_ADDRESS_IN_USE","type":"int128","access":"constant"},{"name":"ERR_STACKING_STX_LOCKED","type":"int128","access":"constant"},{"name":"ERR_STACKING_THRESHOLD_NOT_MET","type":"int128","access":"constant"},{"name":"ERR_STACKING_UNREACHABLE","type":"int128","access":"constant"},{"name":"ERR_STACK_EXTEND_NOT_LOCKED","type":"int128","access":"constant"},{"name":"ERR_STACK_INCREASE_NOT_LOCKED","type":"int128","access":"constant"},{"name":"MAX_ADDRESS_VERSION","type":"uint128","access":"constant"},{"name":"MAX_ADDRESS_VERSION_BUFF_20","type":"uint128","access":"constant"},{"name":"MAX_ADDRESS_VERSION_BUFF_32","type":"uint128","access":"constant"},{"name":"MAX_POX_REWARD_CYCLES","type":"uint128","access":"constant"},{"name":"MIN_POX_REWARD_CYCLES","type":"uint128","access":"constant"},{"name":"POX_REJECTION_FRACTION","type":"uint128","access":"constant"},{"name":"PREPARE_CYCLE_LENGTH","type":"uint128","access":"constant"},{"name":"REWARD_CYCLE_LENGTH","type":"uint128","access":"constant"},{"name":"STACKING_THRESHOLD_100","type":"uint128","access":"constant"},{"name":"STACKING_THRESHOLD_25","type":"uint128","access":"constant"},{"name":"configured","type":"bool","access":"variable"},{"name":"first-2-1-reward-cycle","type":"uint128","access":"variable"},{"name":"first-burnchain-block-height","type":"uint128","access":"variable"},{"name":"pox-prepare-cycle-length","type":"uint128","access":"variable"},{"name":"pox-rejection-fraction","type":"uint128","access":"variable"},{"name":"pox-reward-cycle-length","type":"uint128","access":"variable"}],"maps":[{"name":"allowance-contract-callers","key":{"tuple":[{"name":"contract-caller","type":"principal"},{"name":"sender","type":"principal"}]},"value":{"tuple":[{"name":"until-burn-ht","type":{"optional":"uint128"}}]}},{"name":"delegation-state","key":{"tuple":[{"name":"stacker","type":"principal"}]},"value":{"tuple":[{"name":"amount-ustx","type":"uint128"},{"name":"delegated-to","type":"principal"},{"name":"pox-addr","type":{"optional":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}}},{"name":"until-burn-ht","type":{"optional":"uint128"}}]}},{"name":"logged-partial-stacked-by-cycle","key":{"tuple":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"},{"name":"sender","type":"principal"}]},"value":{"tuple":[{"name":"stacked-amount","type":"uint128"}]}},{"name":"partial-stacked-by-cycle","key":{"tuple":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-cycle","type":"uint128"},{"name":"sender","type":"principal"}]},"value":{"tuple":[{"name":"stacked-amount","type":"uint128"}]}},{"name":"reward-cycle-pox-address-list","key":{"tuple":[{"name":"index","type":"uint128"},{"name":"reward-cycle","type":"uint128"}]},"value":{"tuple":[{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"stacker","type":{"optional":"principal"}},{"name":"total-ustx","type":"uint128"}]}},{"name":"reward-cycle-pox-address-list-len","key":{"tuple":[{"name":"reward-cycle","type":"uint128"}]},"value":{"tuple":[{"name":"len","type":"uint128"}]}},{"name":"reward-cycle-total-stacked","key":{"tuple":[{"name":"reward-cycle","type":"uint128"}]},"value":{"tuple":[{"name":"total-ustx","type":"uint128"}]}},{"name":"stacking-rejection","key":{"tuple":[{"name":"reward-cycle","type":"uint128"}]},"value":{"tuple":[{"name":"amount","type":"uint128"}]}},{"name":"stacking-rejectors","key":{"tuple":[{"name":"reward-cycle","type":"uint128"},{"name":"stacker","type":"principal"}]},"value":{"tuple":[{"name":"amount","type":"uint128"}]}},{"name":"stacking-state","key":{"tuple":[{"name":"stacker","type":"principal"}]},"value":{"tuple":[{"name":"delegated-to","type":{"optional":"principal"}},{"name":"first-reward-cycle","type":"uint128"},{"name":"lock-period","type":"uint128"},{"name":"pox-addr","type":{"tuple":[{"name":"hashbytes","type":{"buffer":{"length":32}}},{"name":"version","type":{"buffer":{"length":1}}}]}},{"name":"reward-set-indexes","type":{"list":{"type":"uint128","length":12}}}]}}],"fungible_tokens":[],"non_fungible_tokens":[],"epoch":"Epoch24","clarity_version":"Clarity2"}`;
 
 // === MOCK MAPS ===============================================================
-export const MOCK_POX_2_REGTEST = {
+export const MOCK_POX_3_REGTEST = {
   '/v2/fees/transaction': V2_FEES,
-  '/v2/pox': V2_POX_REGTEST,
-  '/v2/contracts/interface/ST000000000000000000002AMW42H/pox-2': V2_POX_INTERFACE_POX_2,
-  '/v2/data_var/ST000000000000000000002AMW42H/pox-2/configured?proof=0': `{"data":"0x03"}`,
+  '/v2/pox': V2_POX_REGTEST_POX_3,
+  '/v2/contracts/interface/ST000000000000000000002AMW42H/pox-3': V2_POX_INTERFACE_POX_3,
 };
 
 export const MOCK_EMPTY_ACCOUNT = {
-  '/v2/pox': V2_POX_STACKS_TEL,
+  '/v2/pox': V2_POX_REGTEST_POX_3,
   '/v2/accounts/ST162GBCTD9ESBF09XC2T63NCX6ZKS42ZPWGXZ6VH?proof=0': `{"balance":"0x00000000000000000000000000000000","locked":"0x00000000000000000000000000000000","unlock_height":0,"nonce":0}`,
-  '/v2/contracts/call-read/ST000000000000000000002AMW42H/pox-2/can-stack-stx': `{"okay":true,"result":"0x08000000000000000000000000000000000b"}`,
+  '/v2/contracts/call-read/ST000000000000000000002AMW42H/pox-3/can-stack-stx': `{"okay":true,"result":"0x08000000000000000000000000000000000b"}`,
 };
 
 export const MOCK_FULL_ACCOUNT = {
-  '/v2/pox': V2_POX_STACKS_TEL,
-  '/v2/accounts/STB44HYPYAT2BB2QE513NSP81HTMYWBJP02HPGK6?proof=0': `{"balance":"0x0000000000000000002386f26fb76310","locked":"0x00000000000000000000000000000000","unlock_height":0,"nonce":63}`,
-  '/v2/contracts/call-read/ST000000000000000000002AMW42H/pox-2/can-stack-stx': `{"okay":true,"result":"0x0703"}`,
+  '/v2/pox': V2_POX_REGTEST_POX_3,
+  '/v2/accounts/STB44HYPYAT2BB2QE513NSP81HTMYWBJP02HPGK6?proof=0': `{"balance":"0x0000000000000000001cdd78bdeadfe8","locked":"0x00000000000000000006a979b1d61c00","unlock_height":435,"nonce":4}`,
+  '/v2/contracts/call-read/ST000000000000000000002AMW42H/pox-3/can-stack-stx': `{"okay":true,"result":"0x0703"}`,
 };
 
 // === FETCHMOCK TESTING HELPERS ===============================================
