+++ packages/stacking/tests/utils.test.ts
@@ -11,7 +11,7 @@ export const BTC_ADDRESS_CASES = [
     address: '132hGUmTRtYQFq7gHVsBCfbwFp4eNsQvBi',
     expectedVersion: 0,
     expectedLength: 20,
-    mockedResult: `0x0a0c000000041266697273742d7265776172642d6379636c650100000000000000000000000000001a200b6c6f636b2d706572696f64010000000000000000000000000000000208706f782d616464720c00000002096861736862797465730200000014164247d6f2b425ac5771423ae6c80c754f7172b00776657273696f6e020000000100127265776172642d7365742d696e64657865730b0000000201000000000000000000000000000000000100000000000000000000000000000000`,
+    mockedResult: `0x0a0c000000050c64656c6567617465642d746f091266697273742d7265776172642d6379636c6501000000000000000000000000000000330b6c6f636b2d706572696f64010000000000000000000000000000000108706f782d616464720c00000002096861736862797465730200000014164247d6f2b425ac5771423ae6c80c754f7172b00776657273696f6e020000000100127265776172642d7365742d696e64657865730b000000010100000000000000000000000000000000`,
   },
   {
     format: 'p2sh',
@@ -19,7 +19,7 @@ export const BTC_ADDRESS_CASES = [
     address: '34puHau4ZYdxJrUwy3eYT6ySpwQJxx4rkS',
     expectedVersion: 1,
     expectedLength: 20,
-    mockedResult: `0x0a0c000000041266697273742d7265776172642d6379636c650100000000000000000000000000001a260b6c6f636b2d706572696f64010000000000000000000000000000000208706f782d616464720c0000000209686173686279746573020000001422662a21bb732892ac097e901eb876eec9a7fc4b0776657273696f6e020000000101127265776172642d7365742d696e64657865730b0000000201000000000000000000000000000000000100000000000000000000000000000000`,
+    mockedResult: `0x0a0c000000050c64656c6567617465642d746f091266697273742d7265776172642d6379636c6501000000000000000000000000000000350b6c6f636b2d706572696f64010000000000000000000000000000000108706f782d616464720c0000000209686173686279746573020000001422662a21bb732892ac097e901eb876eec9a7fc4b0776657273696f6e020000000101127265776172642d7365742d696e64657865730b000000010100000000000000000000000000000000`,
   },
   {
     format: 'p2sh-p2wpkh',
@@ -27,7 +27,7 @@ export const BTC_ADDRESS_CASES = [
     address: '3P2pbdfGvBSMuSZZFBartXTtngAzVVGm1N',
     expectedVersion: 1, // 'p2sh-p2wpkh' can't be detected
     expectedLength: 20,
-    mockedResult: `0x0a0c000000041266697273742d7265776172642d6379636c650100000000000000000000000000001a290b6c6f636b2d706572696f64010000000000000000000000000000000208706f782d616464720c00000002096861736862797465730200000014ea198195595aa5ea4b9dd5b3d84cb0387be6ca770776657273696f6e020000000101127265776172642d7365742d696e64657865730b0000000201000000000000000000000000000000000100000000000000000000000000000000`,
+    mockedResult: `0x0a0c000000050c64656c6567617465642d746f091266697273742d7265776172642d6379636c6501000000000000000000000000000000370b6c6f636b2d706572696f64010000000000000000000000000000000108706f782d616464720c00000002096861736862797465730200000014ea198195595aa5ea4b9dd5b3d84cb0387be6ca770776657273696f6e020000000101127265776172642d7365742d696e64657865730b000000010100000000000000000000000000000000`,
   },
   {
     format: 'p2sh-p2wsh',
@@ -35,7 +35,7 @@ export const BTC_ADDRESS_CASES = [
     address: '358ULjuieeSJFfKuwhL4Dk6RETiu85jKN6',
     expectedVersion: 1, // 'p2sh-p2wsh' can't be detected
     expectedLength: 20,
-    mockedResult: `0x0a0c000000041266697273742d7265776172642d6379636c650100000000000000000000000000001a450b6c6f636b2d706572696f64010000000000000000000000000000000208706f782d616464720c0000000209686173686279746573020000001425b8d40efb9428c2a49a1a93ae92e32605347c020776657273696f6e020000000101127265776172642d7365742d696e64657865730b0000000201000000000000000000000000000000000100000000000000000000000000000000`,
+    mockedResult: `0x0a0c000000050c64656c6567617465642d746f091266697273742d7265776172642d6379636c6501000000000000000000000000000000390b6c6f636b2d706572696f64010000000000000000000000000000000108706f782d616464720c0000000209686173686279746573020000001425b8d40efb9428c2a49a1a93ae92e32605347c020776657273696f6e020000000101127265776172642d7365742d696e64657865730b000000010100000000000000000000000000000000`,
   },
   {
     format: 'p2wpkh',
@@ -43,7 +43,7 @@ export const BTC_ADDRESS_CASES = [
     address: 'bc1qzepy04hjksj6c4m3ggawdjqvw48hzu4sy24ghc',
     expectedVersion: 4,
     expectedLength: 20,
-    mockedResult: `0x0a0c000000041266697273742d7265776172642d6379636c650100000000000000000000000000001a490b6c6f636b2d706572696f64010000000000000000000000000000000208706f782d616464720c00000002096861736862797465730200000014164247d6f2b425ac5771423ae6c80c754f7172b00776657273696f6e020000000104127265776172642d7365742d696e64657865730b0000000201000000000000000000000000000000000100000000000000000000000000000000`,
+    mockedResult: `0x0a0c000000050c64656c6567617465642d746f091266697273742d7265776172642d6379636c65010000000000000000000000000000003b0b6c6f636b2d706572696f64010000000000000000000000000000000108706f782d616464720c00000002096861736862797465730200000014164247d6f2b425ac5771423ae6c80c754f7172b00776657273696f6e020000000104127265776172642d7365742d696e64657865730b000000010100000000000000000000000000000000`,
   },
   {
     format: 'p2wsh',
@@ -51,7 +51,7 @@ export const BTC_ADDRESS_CASES = [
     address: 'bc1qczxqnu5hx2zvxcqt3lmr6vjju4ysf7d649mvrzd8v7l3jez0dqzqgz0ewt',
     expectedVersion: 5,
     expectedLength: 32,
-    mockedResult: `0x0a0c000000041266697273742d7265776172642d6379636c650100000000000000000000000000001a9c0b6c6f636b2d706572696f64010000000000000000000000000000000208706f782d616464720c00000002096861736862797465730200000020c08c09f2973284c3600b8ff63d3252e54904f9baa976c189a767bf19644f68040776657273696f6e020000000105127265776172642d7365742d696e64657865730b0000000201000000000000000000000000000000000100000000000000000000000000000000`,
+    mockedResult: `0x0a0c000000050c64656c6567617465642d746f091266697273742d7265776172642d6379636c65010000000000000000000000000000003d0b6c6f636b2d706572696f64010000000000000000000000000000000108706f782d616464720c00000002096861736862797465730200000020c08c09f2973284c3600b8ff63d3252e54904f9baa976c189a767bf19644f68040776657273696f6e020000000105127265776172642d7365742d696e64657865730b000000010100000000000000000000000000000000`,
   },
   {
     format: 'p2tr',
@@ -59,7 +59,7 @@ export const BTC_ADDRESS_CASES = [
     address: 'bc1p8dlmzllfah294ntwatr8j5uuvcj7yg0dete94ck2krrk0ka2c9qqwwn4dr',
     expectedVersion: 6,
     expectedLength: 32,
-    mockedResult: `0x0a0c000000041266697273742d7265776172642d6379636c650100000000000000000000000000001a9f0b6c6f636b2d706572696f64010000000000000000000000000000000208706f782d616464720c000000020968617368627974657302000000203b7fb17fe9edd45acd6eeac679539c6625e221edcaf25ae2cab0c767dbaac1400776657273696f6e020000000106127265776172642d7365742d696e64657865730b0000000201000000000000000000000000000000000100000000000000000000000000000000`,
+    mockedResult: `0x0a0c000000050c64656c6567617465642d746f091266697273742d7265776172642d6379636c65010000000000000000000000000000003f0b6c6f636b2d706572696f64010000000000000000000000000000000108706f782d616464720c000000020968617368627974657302000000203b7fb17fe9edd45acd6eeac679539c6625e221edcaf25ae2cab0c767dbaac1400776657273696f6e020000000106127265776172642d7365742d696e64657865730b000000010100000000000000000000000000000000`,
   },
 ];
